Fixed tweets not opening when a profile, search, or tweet screen is already open.
When launching Fritter from a Twitter URI in another application, if
Fritter is already open, the link appears to not open, as it replaced
a widget underneath the currently-shown one.

This fixes the issue by always pushing a new navigation item on top of
what's already displayed. It does mean the navigation stack gets longer
(i.e. pressing back will progress through the stack instead of going
back to the calling app), but that's probably preferable to nothing
showing at all 😅.

Fixes #131.