Improved the layout of the video player for narrower screens.
This upgrades `chewie` to the latest version, and imports the new
`MaterialDesktopControls`, which are a bit more user friendly, and
small screen friendly.

The progress bar is now displayed above the controls, allowing it to be
seen on narrower device screens in portrait mode, and it also moves the
player actions to a menu.

Fixes #127.