Added support for automatic searching when text is entered.
This adds support for automatically searching Twitter when a query is
entered on the search screen. It also means that tapping on a hashtag
actually performs the search now, instead of having to manually press
search 🙌.

The search request is debounced now too, so it doesn't perform a search
any more often than every 750ms, meaning we're not spamming Twitter's
API with a search for every character typed in.

Fixes #25 and #31.