Added verified icon to the user tile.
This adds the verified icon next to a verified user's name on the search
and subscription screens.

It adds a new column to the `TABLE_SUBSCRIPTIONS` table, so there's a
new migration. It also means that for the icon to appear on the
subscriptions screen, a manual refresh is required.

Fixes #121.