Added Sentry as the opt-in error reporting location
This change replaces the email error reporting function with an opt-in
reporting function that sends the error to Sentry.

If the user has set no preference for error reporting, a dialog is
shown, giving the user 4 options:

* Send once: report the error just this one time
* Send always: always report errors to Sentry, automatically
* Don't send: don't report the error this time
* Never send: never reports to Sentry, ever

The preference is also exposed in the options screen, giving the user
total control on whether error reporting to Sentry is enabled or not.