Now disposing pages and tweets
This is intended to help with memory usage, by disposing
PagingController instances after a view is closed, and also by
introducing automatic keepalives on tweets so they get disposed when
they get far enough off screen.

These changes didn't make much of a difference on my device, but they
should help in theory.

This hopefully addresses #182, but I'm not saying it fixes it just yet.