Improved handling of unexpected responses from Twitter
This improves the displaying and reporting of errored HTTP responses
from Twitter, e.g. when some parameters are incorrect.

Currently they just show as "Instance of 'Response'" which isn't helpful
to the user or developers, and now they report the status code, reason
phrase, error body and request URI.

This should help solve a few open issues, including #82 and #193.