Fixed an overflow that was causing feeds to error for some users
This fixes a long-standing bug where the logic that paginates feeds with
large numbers of members wasn't correctly limiting the API query length,
causing Twitter to return an "invalid parameters" message.

This would only reproduce with very specific members in groups, as it
needs just one long account name to cause the string to overflow.

I've lowered the string length that forces a new API request, so this
shouldn't happen anymore.

Fixes #82 and #193.