Fixed feed scroll position being reset when going back. Fixes #237
This fixes a regression made in v2.14.0 where, on the Feed tab of the
home screen, selecting any tweet and then going back (after any depth of
page navigation) causes the Feed tab's content to refresh, and lose the
scroll position, returning to the top.

This was caused by creating a new instance of the FeedScreen component
every time the home screen was rendered, which was implemented to
achieve "scroll to top" functionality, so I fixed it by returning back
to initializing it once, but when the home screen is first created
instead of in a field.