Tweets are now properly disposed, freeing up memory
This was attempted in 20fe16, but I misread the documentation and that
resulted in no change whatsoever. Now, tweets that are displayed in a
paged list (i.e. in the feed, a group, or on a profile) are now disposed
when they get far enough out of the viewport, freeing up memory and
improving performance.

This also appears to solve a lot of video issues, as ExoPlayer can only
have a finite number of instances in memory (or something along those
lines), and now those are also being disposed properly. So, I'll say
this solve #198 and #161 too.

This solves #182 to a large extent.