Fixed infinite follower loading (#283)
This was being caused by a missing check in the paginator, where the
last page returns "0" for the pagination cursor, but we were continuing
to paginate due to Twitter returning "-1" on the next page, making them
not match.

I fixed this by adding a branch to append the last page if we ever get a
bottom cursor with the value "0".