Fixed some characters being unsupported when searching for users (#209)
This seems like it was caused by a limitation in the Twitter API
endpoint that was being used (/1.1/users/search.json).

I fixed it by swapping to the adaptive.json endpoint, which seems to be
a bit richer, and supports unicode characters. It also supports
pagination, which should be useful in the future.