Fixed error being displayed when doubly-quoted tweets are encountered
This was being caused by Fritter trying to emulate the Twitter web
interface, where if a nested tweets can't be found, it displays a "tweet
unavailable" message, but it missed the case where a quoted or retweeted
tweet itself has a quoted or retweeted tweet, which the Twitter web
interface actually ignores instead of displaying an error.

This fixes #397.